Change file structure of DomainMessages.
